MeetingAssistant/requirements.txt
gradio>=4.44.0
openai-whisper>=20231117
torch>=2.1.0
ollama>=0.4.0